app/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.46.0",
    "numpy>=2.2.3",
    "openai>=1.63.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "polygon>=1.2.6",
    "polygon-api-client>=1.14.3",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.42.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
    "yfinance>=0.2.53",
    "python-chess>=1.999",
]
